1. Item-based CF with jaccard similarity matrix based on LOD properties 2. Improved personalized PageRank with Jaccard similarity

+++ src/main/java/di/uniba/it/lodrecsys/graph/UserItemJaccardScore.java
@@ -169,7 +169,7 @@ public class UserItemJaccardScore extends RecGraph {
             String resourceURI = idUriMap.get(currItemID);
             Double finalScore = priors.getVertexScore(currItemID); // pageRankScore
             if (resourceURI != null) {
-                Double jaccardScore = calculator.getSimilarity().compute(userCentroid, loadItemRepresentation(resourceURI));
+                Double jaccardScore = (double) calculator.getSimilarity().compute(userCentroid, loadItemRepresentation(resourceURI));
                 finalScore = (2 * finalScore * jaccardScore) / (jaccardScore + finalScore);
 
             }
